feat: v0.2.0 — full PI tracker feature set
Characters, Extractors, Planets, Projects, Corp PI, Optimizer, Profit
calculator, Build-out, Maintenance log, Market price sync via Fuzzwork.

ESI sync via django-esi ESIClientProvider (ETag/304, error-limit
handling, QueueOnce dedup, autoretry). German translation included.

- Rename fetchPost → piPost to avoid collision with AA auth-framework.js
- Add @ensure_csrf_cookie to characters_page; gate share button on perm
- Fix corp optimizer using personal characters when no corp pool existed
- Add @require_POST to all JSON-only views; remove manual method checks
- Bump django-esi requirement to >=9.0.0 (ESIClientProvider)
- Add project.urls to pyproject.toml; add esi-skills scope to README
- Remove useless GeneralAdmin from admin.py
- Profit page: show direct recipe inputs instead of P0 expansion; POCO default 0
- Drop Cycle and P0-implied columns from profit table

+++ aa_pi_tracker/models/settings.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class PiUserSettings(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name="pi_settings")
-    home_system_id = models.IntegerField(null=True, blank=True)
+    home_system_id = models.BigIntegerField(null=True, blank=True)
     home_system_name = models.CharField(max_length=100, blank=True, default="")
 
     class Meta:
